requirements.txt
flask==3.0.3
gunicorn==21.2.0
moviepy==1.0.3
imageio-ffmpeg==0.4.9
Pillow==10.4.0
numpy==1.26.4